% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.secure_result}
\alias{print.secure_result}
\alias{summary.secure_result}
\title{Print a Secure Result}
\usage{
\method{print}{secure_result}(x, ...)

\method{summary}{secure_result}(object, ...)
}
\arguments{
\item{x}{A \code{secure_result} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a Secure Result
}
