% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{print.shieldr_report}
\alias{print.shieldr_report}
\title{Print a \code{shieldr_report}}
\usage{
\method{print}{shieldr_report}(x, ..., show_stats = FALSE)
}
\arguments{
\item{x}{A \code{shieldr_report}.}

\item{...}{Unused.}

\item{show_stats}{Show formatting time, network status, and report token
count when available.}
}
\value{
The report, invisibly.
}
\description{
Print a \code{shieldr_report}
}
