% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{print.shieldr_report}
\alias{print.shieldr_report}
\title{Print a \code{shieldr_report}}
\usage{
\method{print}{shieldr_report}(x, ...)
}
\arguments{
\item{x}{A \code{shieldr_report}.}

\item{...}{Unused.}
}
\value{
The report, invisibly.
}
\description{
Print a \code{shieldr_report}
}
